@@ -41,7 +41,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; 2016 - Sandbox</p>
+            <p>&copy; @DateTime.Now.Year :: Sandbox :: Powered by @System.Runtime.InteropServices.RuntimeInformation.FrameworkDescription</p>
         </footer>
     </div>
 
